🐛 Ignore UdpSocket's WSAECONNRESET error on Windows
